Change type to int64_t

+++ evalAssignment1/update_locations.c
@@ -1,20 +1,20 @@
 #include "update_locations.h"
 
-// Create a list of 'size' int16_ting point numbers in the range [-bound, bound]
-int16_t * generate_random_list(int32_t size, int32_t bound) {
-    int16_t* array;
-    array = (int16_t*)malloc(sizeof(int16_t) * size);
+// Create a list of 'size' int64_ting point numbers in the range [-bound, bound]
+int64_t * generate_random_list(int32_t size, int32_t bound) {
+    int64_t* array;
+    array = (int64_t*)malloc(sizeof(int64_t) * size);
 
 
     for (int i = 0; i < size; i++) {
-        int16_t num = (int16_t)(rand() % (2*bound+1)) - bound;
+        int64_t num = (int64_t)(rand() % (2*bound+1)) - bound;
         array[i] = num;
     }
     return array;
 }
 
 // Update position by velocity, one time-step
-void update_coords(int16_t * xs, int16_t * ys, int16_t * zs, int16_t * vx, int16_t * vy, int16_t * vz, int32_t size){
+void update_coords(int64_t * xs, int64_t * ys, int64_t * zs, int64_t * vx, int64_t * vy, int64_t * vz, int32_t size){
     for (int i = 0; i < size; i ++) {
         xs[i] = xs[i] + vx[i];
         ys[i] = ys[i] + vy[i];
@@ -22,8 +22,8 @@ void update_coords(int16_t * xs, int16_t * ys, int16_t * zs, int16_t * vx, int16
     }
 }
 
-int16_t sum(int16_t * array, int32_t size) {
-    int16_t sum = 0;
+int64_t sum(int64_t * array, int32_t size) {
+    int64_t sum = 0;
     for (int i = 0; i < size; i++) { sum += array[i];}
     return sum;
 }
@@ -37,12 +37,12 @@ int main(int argc, char* argv[]) {
     int32_t size = atoi(argv[1]);
     int32_t iters = atoi(argv[2]);
     srand(size);
-    int16_t * xs = generate_random_list(size, 1000);
-    int16_t * ys = generate_random_list(size, 1000);
-    int16_t * zs = generate_random_list(size, 1000);
-    int16_t * vx = generate_random_list(size, 1);
-    int16_t * vy = generate_random_list(size, 1);
-    int16_t * vz = generate_random_list(size, 1);
+    int64_t * xs = generate_random_list(size, 1000);
+    int64_t * ys = generate_random_list(size, 1000);
+    int64_t * zs = generate_random_list(size, 1000);
+    int64_t * vx = generate_random_list(size, 1);
+    int64_t * vy = generate_random_list(size, 1);
+    int64_t * vz = generate_random_list(size, 1);
 
     struct timespec start, stop;
     clock_gettime(CLOCK_MONOTONIC, &start);
@@ -50,14 +50,14 @@ int main(int argc, char* argv[]) {
         update_coords(xs, ys, zs, vx, vy, vz, size);
     }
     clock_gettime(CLOCK_MONOTONIC, &stop);
-    int16_t chksum = sum(xs, size) + sum(ys, size) + sum(zs, size);
+    int64_t chksum = sum(xs, size) + sum(ys, size) + sum(zs, size);
     double obj_count = (double)size*(double)iters;
     double time_taken;
     time_taken = ((double)stop.tv_sec - (double)start.tv_sec) * 1e9;
     time_taken = ((double)time_taken + ((double)stop.tv_nsec - start.tv_nsec)) * 1e-9;
     time_taken = time_taken * 1e6 / obj_count;
     printf("%fus\n",time_taken);
-    printf("Final checksum is: %d", chksum);
+    printf("Final checksum is: %lld", chksum);
 
     free(xs);
     free(ys);
